DespensaClassV3

+++ app/src/main/java/com/ufabc/kleinzanin/homemarket/adapter/DespensaAdapterAdd.java
@@ -72,10 +72,10 @@ public class DespensaAdapterAdd extends BaseAdapter {
         quantidade = (TextView ) convertView.findViewById(R.id.produto_quantidade);
         nome.setText(produto.getNome());
         if(produto.getChecked() == true){
-            quantidade.setText(Integer.toString(produto.getQuantidade()) + "/" +
-                    Integer.toString(produto.getConsumo()) + " " + produto.getUnidade());}
+            quantidade.setText(Double.toString(produto.getQuantidade()) + "/" +
+                    Double.toString(produto.getConsumo()) + " " + produto.getUnidade());}
         else{
-            quantidade.setText(Integer.toString(produto.getQuantidade()) + " " + produto.getUnidade());
+            quantidade.setText(Double.toString(produto.getQuantidade()) + " " + produto.getUnidade());
 
         }
         return convertView;
