+++ app/src/main/java/com/ufabc/kleinzanin/homemarket/adapter/ProdutoAdapter.java
@@ -74,10 +74,10 @@ public class ProdutoAdapter extends BaseAdapter {
         quantidade = (TextView ) convertView.findViewById(R.id.produto_quantidade);
         nome.setText(produto.getNome());
         if(produto.getChecked() == true){
-        quantidade.setText(Integer.toString(produto.getQuantidade()) + "/" +
-                Integer.toString(produto.getConsumo()) + " " + produto.getUnidade());}
+        quantidade.setText(Double.toString(produto.getQuantidade()) + "/" +
+                Double.toString(produto.getConsumo()) + " " + produto.getUnidade());}
         else{
-            quantidade.setText(Integer.toString(produto.getQuantidade()));
+            quantidade.setText(Double.toString(produto.getQuantidade()));
 
     }
         return convertView;
